🏗️ 페이지 서버/클라이언트 컴포넌트 구조 분리 적용
- 페이지의 진입점을 서버 컴포넌트로 변경하여 useSearchParams Suspense 이슈 해결
- 클라이언트 로직을 독립적인 Client 컴포넌트로 추출 및 배치
